Change from SUSET to USERSET
It doesn't leak any info about the blocking process, so no reason for
it to be SUSET.

It was trying to pass some additional info, but that data never made it
into the sent NOTICE, so stop trying to include it.  It would be nice
for a superuser to get the text of the blocking statement, but it is not
clear how to do that.