Refactor set-up
This often finds problems which don't need crashes at all to reproduce.
Move the crash-specific settings up top so they can be set to empty
string. That way this can be run against an unpatched server more easily.

Also, make the setting of overly-verbose guc parameters less silly, by
not turning them on rather than turning them and back off again.  This
is required for JJ_vac, as it doesn't exist for unpatched servers and is
an error to turn it on even trasiently.