Show rusage information
When log_min_duration_statement is exceeded, show the user and system
CPU time in addition to the wall-clock time.

This is useful to know if a random slow statement might have been
due to transient IO problems or needing to share the CPU with too many
other processes, or was just inherently slow.

It should probably avoid calling both GetCurrentTimestamp
and pg_rusage_init at the start and stop of each statment, as the
timestamp should already be part of the pg_rusage_init.  But that
would be harder to code, and the overhead is not high enough to justify
it for a quick hack.