Rearrange locking in WAL sync
Make competing backends block on the fdatasync itself rather than on the
WALWriteLock under control of JJFsyncLock.  The theory is that while one
process is waiting for the platter to get around so its commit record can
be written, another process can write more WAL, and the kernel might be
smart enough to realize it can combine the writes into one.  But
apparently it is not smart enough to do that, no improvement was detected.

If this had panned out, perhapse the post-sync work could have been arranged
to happen under a spinlock, so that WALWriteLock did not need to be taken
again after the sync.