Add ability to turn off WAL logging.
This is useful for assessing where the bottlenecks are. Obviously
it is not crash-safe.

It still emits some log messages which are needed for bootstrap,
clean shutdown, and restart.  I don't know exactly which messages
those are, so just kind of guessed.  FPI_FOR_HINT occurs a lot
and doesn't seem to be vital, so I specially excluded them.

Using pgbench with fsync off on a AWS 72 core instance (5.18xlarge)
shows that the WAL insertion is a big bottleneck under high
concurrency, turning off WAL increased throughput 2.5 fold.